+++ op/k8s/proxy_boot.go
@@ -4,10 +4,11 @@ import (
 	"context"
 	"fmt"
 
+	"k8s.io/client-go/tools/clientcmd"
+
 	"github.com/cybozu-go/cke"
 	"github.com/cybozu-go/cke/op"
 	"github.com/cybozu-go/cke/op/common"
-	"k8s.io/client-go/tools/clientcmd"
 )
 
 const (