+++ op/nodedns/nodedns.go
@@ -4,9 +4,10 @@ import (
 	"bytes"
 	"text/template"
 
-	"github.com/cybozu-go/cke/op"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/cybozu-go/cke/op"
 )
 
 type unboundConfigTemplate struct {