+++ pkg/ckecli/cmd/repair_queue_add.go
@@ -3,9 +3,10 @@ package cmd
 import (
 	"context"
 
-	"github.com/cybozu-go/cke"
 	"github.com/cybozu-go/well"
 	"github.com/spf13/cobra"
+
+	"github.com/cybozu-go/cke"
 )
 
 var repairQueueAddCmd = &cobra.Command{