+++ sabakan/query_test.go
@@ -6,9 +6,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cybozu-go/cke/sabakan/mock"
 	"github.com/cybozu-go/well"
 	"github.com/google/go-cmp/cmp"
+
+	"github.com/cybozu-go/cke/sabakan/mock"
 )
 
 func testMachine(t *testing.T, m Machine) {