+++ sabakan/repairer.go
@@ -3,8 +3,9 @@ package sabakan
 import (
 	"strings"
 
-	"github.com/cybozu-go/cke"
 	"github.com/cybozu-go/log"
+
+	"github.com/cybozu-go/cke"
 )
 
 func Repairer(machines []Machine, repairEntries []*cke.RepairQueueEntry, rebootEntries []*cke.RebootQueueEntry, nodeStatuses map[string]*cke.NodeStatus, constraints *cke.Constraints) []*cke.RepairQueueEntry {
@@ -24,7 +25,7 @@ func Repairer(machines []Machine, repairEntries []*cke.RepairQueueEntry, rebootE
 	newMachines := make([]Machine, 0, len(machines))
 	for _, machine := range machines {
 		if len(machine.Spec.IPv4) == 0 {
-			log.Warn("ignore non-healthy machine w/o IPv4 address", map[string]interface{}{
+			log.Warn("ignore non-healthy machine w/o IPv4 address", map[string]any{
 				"serial": machine.Spec.Serial,
 			})
 			continue
@@ -34,7 +35,7 @@ func Repairer(machines []Machine, repairEntries []*cke.RepairQueueEntry, rebootE
 		address := machine.Spec.IPv4[0]
 
 		if recent[address] {
-			log.Warn("ignore recently-repaired non-healthy machine", map[string]interface{}{
+			log.Warn("ignore recently-repaired non-healthy machine", map[string]any{
 				"serial":  serial,
 				"address": address,
 			})
@@ -43,7 +44,7 @@ func Repairer(machines []Machine, repairEntries []*cke.RepairQueueEntry, rebootE
 
 		if machine.Status.State == StateUnreachable && machine.Status.Duration < float64(constraints.RepairRebootingSeconds) {
 			if rebooting[address] {
-				log.Info("ignore rebooting unreachable machine", map[string]interface{}{
+				log.Info("ignore rebooting unreachable machine", map[string]any{
 					"serial":  serial,
 					"address": address,
 				})
@@ -51,7 +52,7 @@ func Repairer(machines []Machine, repairEntries []*cke.RepairQueueEntry, rebootE
 			}
 
 			if _, ok := nodeStatuses[address]; !ok {
-				log.Info("ignore out-of-cluster unreachable machine", map[string]interface{}{
+				log.Info("ignore out-of-cluster unreachable machine", map[string]any{
 					"serial":  serial,
 					"address": address,
 				})
@@ -74,7 +75,7 @@ func Repairer(machines []Machine, repairEntries []*cke.RepairQueueEntry, rebootE
 		address := machine.Spec.IPv4[0]
 		serial := machine.Spec.Serial
 		entry := cke.NewRepairQueueEntry(operation, typ, address, serial)
-		log.Info("initiate sabakan-triggered automatic repair", map[string]interface{}{
+		log.Info("initiate sabakan-triggered automatic repair", map[string]any{
 			"serial":    machine.Spec.Serial,
 			"address":   address,
 			"operation": operation,